refactor(users-signals): move signals to own file
